### 2015-04-08 #9 @cboettig
packages repo
- 1. What does this package do? (explain in 50 words or less)

Provides a CRAN-like repository for serving our R packages. See https://github.com/cboettig/packages readme for details.

- 2. Paste the full DESCRIPTION file inside a code block (bounded by ``` on either end).

N/A

- 3. URL for the package (the development repository, not a stylized html page)

https://github.com/cboettig/packages

- 4. What data source(s) does it work with (if applicable)?

R packages from https://github.com/ropensci

- 5. Who is the target audience?

Anyone wanting to install our R packages.

- 6. Check the box next to each policy below, confirming that you agree. These are mandatory.

N/A

* [ ] The repository has continuous integration with Travis and/or another service
* [ ] The package contains a vignette
* [ ] The package contains a reasonably complete readme with devtools install instructions
* [ ] The package contains unit tests

N/A

- 7. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.

I agree.

* [x] Are there any packages that your package depends on that are not on CRAN?

N/A

N/A

* [ ] Does the package have a CRAN accepted license?

N/A

* [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.

N/A

- 8. Please add explanations below for any exceptions to the above:

N/A

- 9. If this is a resubmission following rejection, please explain the change in cirucmstances.
