### 2015-04-09 #9 @Robinlovelace
One of the things most urgently needed in CRAN (or its equivalent), as documented in this excellent [paper](http://www.stat.tugraz.at/AJS/ausg121/121Hornik.pdf), is the need to rate and review packages, so people have an idea of whether they are downloading garbage or not. This would be the main selling point for me. Great idea though, love the fact that it will just work with `install.packages`.
