### 2015-04-23 #9 @karthik
@mine-cetinkaya-rundel PS: Reviewer guidelines are here (https://github.com/ropensci/onboarding/wiki/For-Reviewers) but feel free to ping over email with additional questions.
