### 2015-05-13 #9 @cboettig
Closing this for now until the repo can be updated with better scripts from @richfitz
