### 2015-05-29 #9 @sckott
@cboettig where are we at on this?  still waiting on some infusion of @richfitz -ness
