### 2015-05-29 #9 @cboettig
The draft implementation is in https://github.com/cboettig/drat, and is
currently deployed on http://carlboettiger.info/drat via GitHub.  There's
still a few issues I need to work out with @richfitz, a bit more work to be
done on the automated deploy / update via circle, and some general tidying
and documentation.

On Fri, May 29, 2015 at 12:48 PM Scott Chamberlain <notifications@github.com>
wrote:

> @cboettig <https://github.com/cboettig> where are we at on this? still
> waiting on some infusion of @richfitz <https://github.com/richfitz> -ness
>
