### 2015-05-29 #9 @cboettig
yes.  (the repo still works, though clearly that would be a good place to
put up a documentation page.)

this'll move over to an ropensci host anyway -- that's another thing to
work out.  We might host on a private server instead of Github just to
process the download logs....

On Fri, May 29, 2015 at 1:13 PM Scott Chamberlain <notifications@github.com>
wrote:

> Cool, exciting.
>
> is http://www.carlboettiger.info/drat/ supposed to 404?
>
