### 2015-05-29 #9 @cboettig
added these to the issues list https://github.com/cboettig/drat/issues/1

On Fri, May 29, 2015 at 1:16 PM Scott Chamberlain <notifications@github.com>
wrote:

> heh, something for the future
>
