### 2015-06-24 #9 @cboettig
I think we're ready to reopen this, as I think most of the pieces are now in place and we should migrate this to the ropensci accounts sooner than later (which might still be a bit of a pain updating all paths, but I can keep my branch open if necessary).

I've updated the issues on https://github.com/cboettig/drat/issues/1

Still needs some further documentation (e.g. an nice index.html page), the download stats need to be tabulated, and we have to pick a name / url for the S3 bucket (e.g. packages.ropensci.org or drat.ropensci.org or whatever)
