### 2015-10-06 #9 @sckott
@cboettig since this is in ropensci now, should we close this? or do you still want to do a review
