### 2015-10-06 #9 @cboettig
We can close

On Mon, Oct 5, 2015, 10:52 PM Scott Chamberlain <notifications@github.com>
wrote:

> @cboettig <https://github.com/cboettig> since this is in ropensci now,
> should we close this? or do you still want to do a review
>
